+++ app/controllers/site_controller.rb
@@ -3,5 +3,10 @@
 class SiteController < ApplicationController
   def index
     # Just render a view
+    if params[:character]
+      @character = Character.where(id: params[:id], public: true).pluck(:name, :description).first
+    elsif params[:qc]
+      @character = Qc.where(id: params[:id], public: true).pluck(:name, :description).first
+    end
   end
 end